refactor(benchmark): split oversized runner.go and report.go under the 500-line limit (#158)
runner.go (2131) and report.go (1405) far exceeded coding-standard.md §1's
500-line-per-file limit. Pure behavior-preserving reorganization within
package benchmark — functions are relocated verbatim, no signatures or logic
change. Same-package moves cannot break references; only per-file imports
changed (goimports).

runner.go (2131 -> 442) keeps the core Runner type and lifecycle
(New/Register/Run/RunWithHarness, retry, executeWorkload dispatch). Extracted:
- oracle.go       expected-result construction + truth-pass + candidate filtering
- execute.go      service query execution (executeService*, route info)
- execute_query.go query request/condition/options builders + result helpers
- loaded_state.go loaded-state snapshot + hot-record scan
- assertions.go   the validate* family + record matching / pushdown / per-tier

report.go (1405 -> 309) keeps writers + formatters + metadata. Extracted:
- report_types.go the metadata/summary/diff/trend type declarations
- summary.go       SummarizeRunResult, summarize*, Compare*, compareWorkloadSummaries
- trend.go         trend history/analysis (ReadTrendHistory .. AnalyzeTrend)

No non-generated source file in the package now exceeds 500 lines. Top-level
declaration counts are identical before vs after (108 func / 24 type / 4 const
/ 4 var = 140), so nothing was dropped.

Verification: go build ./...; go vet -tags e2e ./internal/e2e_harness/federated/...;
make test; make lint; the e2e guard tests (TestBenchmarkWorkloadExecution_RunWithHarness,
TestBenchmarkTruthPassSampledSpotCheck_RunWithHarness) and the batching
characterization test all pass — proving the reorg is behavior-preserving.

Closes #158.
