+++ internal/e2e_harness/federated/benchmark/runner.go
@@ -2,23 +2,14 @@ package benchmark
 
 import (
 	"context"
-	"database/sql"
 	"fmt"
-	"sort"
-	"strconv"
 	"sync"
 	"time"
 
-	fedengine "github.com/lychee-technology/forma/internal/federated"
 	"github.com/lychee-technology/forma/internal/model"
 	"github.com/lychee-technology/forma/internal/queryplan"
-	"github.com/lychee-technology/forma/internal/schemameta"
-	"github.com/lychee-technology/forma/internal/transform"
 
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5/pgxpool"
 	forma "github.com/lychee-technology/forma"
-	"github.com/lychee-technology/forma/internal"
 	federated "github.com/lychee-technology/forma/internal/e2e_harness/federated"
 )
 
@@ -346,50 +337,6 @@ func (r *Runner) RunWithHarness(ctx context.Context, h *federated.FederatedTestH
 	return result, nil
 }
 
-func (r *Runner) buildExpectedResults(ctx context.Context, h *federated.FederatedTestHarness, loadedRecords []GeneratedRecord, hotKeys map[string]struct{}) (map[string]expectedWorkloadResult, map[string]string, []string, error) {
-	results := buildExpectedWorkloadResultsFromRecords(loadedRecords, r.workloads, r.config.PageSize, r.genConfig, hotKeys)
-	oracleModes := make(map[string]string, len(r.workloads))
-	loadedStateCount := 0
-	truthPassCount := 0
-	sampledCount := 0
-	var sampleNotes []string
-	for _, workload := range r.workloads {
-		mode := string(workload.ResolvedOracleMode())
-		oracleModes[workload.Name] = mode
-		switch workload.ResolvedOracleMode() {
-		case OracleModeTruthPass:
-			expected, stats, err := buildExpectedWorkloadResultFromFederatedTruth(ctx, h, workload, r.config.PageSize, loadedRecords, r.genConfig, r.config.TruthPassSampleCap, r.config.Seed)
-			if err != nil {
-				return nil, nil, nil, fmt.Errorf("build truth-pass expected result for %s: %w", workload.Name, err)
-			}
-			results[workload.Name] = expected
-			truthPassCount++
-			if stats.Applied {
-				oracleModes[workload.Name] = string(OracleModeTruthPassSampled)
-				sampledCount++
-				sampleNotes = append(sampleNotes, fmt.Sprintf("truth_pass_sample workload=%s cap=%d candidates=%d sampled=%d", workload.Name, stats.Cap, stats.Candidates, stats.Sampled))
-			}
-		default:
-			loadedStateCount++
-		}
-	}
-	summary := fmt.Sprintf("oracle_modes loaded_state=%d truth_pass=%d", loadedStateCount, truthPassCount)
-	if sampledCount > 0 {
-		summary = fmt.Sprintf("%s truth_pass_sampled=%d", summary, sampledCount)
-	}
-	return results, oracleModes, append([]string{summary}, sampleNotes...), nil
-}
-
-// resolveRunOracleMode returns the run-time oracle mode for a workload,
-// preferring the (possibly sampled) mode recorded in oracleModes and falling
-// back to the workload's static resolved mode when the map has no entry.
-func resolveRunOracleMode(oracleModes map[string]string, workload WorkloadDefinition) string {
-	if mode, ok := oracleModes[workload.Name]; ok && mode != "" {
-		return mode
-	}
-	return string(workload.ResolvedOracleMode())
-}
-
 func (r *Runner) validateFixtures() error {
 	for _, fixture := range r.schemas {
 		id, _, err := r.registry.GetSchemaAttributeCacheByName(fixture.Name)
@@ -492,1640 +439,4 @@ func (r *Runner) executeWorkload(ctx context.Context, h *federated.FederatedTest
 		}
 	}
 	return run, result.Records, nil
-}
-
-func extractRouteInfo(plan *model.ExecutionPlan) (engine, reason string) {
-	if plan == nil {
-		return "", ""
-	}
-	if plan.Routing.UseDuckDB {
-		engine = "duckdb"
-	} else if len(plan.Routing.Reason) > 0 {
-		engine = "postgres"
-	}
-	reason = plan.Routing.Reason
-	return
-}
-
-func routeInfoFromPlanNotes(notes []string) (engine, reason string) {
-	for _, note := range notes {
-		switch note {
-		case "postgres_only_execution", "prefer_hot_override":
-			engine = "postgres"
-			reason = note
-		case "dirty_ids_excluded":
-			if engine == "" {
-				engine = "duckdb"
-				reason = "federated"
-			}
-		}
-	}
-	if engine == "" && len(notes) > 0 {
-		engine = "duckdb"
-		reason = "federated"
-	}
-	return
-}
-
-func (r *Runner) executeServiceWorkload(ctx context.Context, h *federated.FederatedTestHarness, workload WorkloadDefinition) (WorkloadRunResult, []*model.PersistentRecord, error) {
-	req, pageSize := queryRequestForWorkload(workload, r.config.PageSize)
-	start := time.Now()
-	capturePlan := workload.Category == WorkloadCategoryPushdown || workload.Category == WorkloadCategoryTierMix
-
-	var result *forma.QueryResult
-	var records []*model.PersistentRecord
-	var plan *model.ExecutionPlan
-	var err error
-
-	if workload.UseKeysetPagination {
-		result, records, plan, err = r.executeKeysetServiceQuery(ctx, h, workload)
-	} else {
-		result, records, plan, err = executeServiceQueryWithPlan(ctx, h, req, r.config.PageSize, capturePlan, r.planCache)
-	}
-	if err != nil {
-		return WorkloadRunResult{}, nil, err
-	}
-	run := WorkloadRunResult{
-		Name:         workload.Name,
-		Category:     string(workload.Category),
-		Distribution: r.genConfig.Distribution,
-		PageSize:     pageSize,
-		PageNumber:   workload.PageNumber,
-		Offset:       workload.DerivedOffset(r.config.PageSize),
-		PreferHot:    workload.PreferHot,
-		ResultCount:  len(records),
-		TotalRecords: int64(result.TotalRecords),
-		Duration:     time.Since(start),
-		Passed:       true,
-		RowIDs:       persistentRecordIDs(records),
-		PlanNotes:    []string{"entity_manager_query_service"},
-		PerTier:      extractPerTierMetrics(plan),
-	}
-	if workload.UseKeysetPagination {
-		run.PaginationMode = "keyset"
-		run.PlanNotes = append(run.PlanNotes, "keyset_pagination")
-	}
-	if engine, reason := extractRouteInfo(plan); engine != "" {
-		run.RouteEngine = engine
-		run.RouteReason = reason
-	}
-	return run, records, nil
-}
-
-func (r *Runner) executeKeysetServiceQuery(ctx context.Context, h *federated.FederatedTestHarness, workload WorkloadDefinition) (*forma.QueryResult, []*model.PersistentRecord, *model.ExecutionPlan, error) {
-	if h == nil || h.PGDSN == "" {
-		return nil, nil, nil, fmt.Errorf("benchmark harness postgres DSN is required")
-	}
-	pool, err := pgxpool.New(ctx, h.PGDSN)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("connect benchmark pgx pool: %w", err)
-	}
-	defer pool.Close()
-
-	if err := RegisterFixtureSchemas(h); err != nil {
-		return nil, nil, nil, fmt.Errorf("register fixture schemas: %w", err)
-	}
-	schemaTable, err := ensureBenchmarkSchemaRegistry(ctx, pool)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("prepare benchmark schema registry: %w", err)
-	}
-
-	registry, err := schemameta.NewFileSchemaRegistry(pool, schemaTable, FixturesDir())
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("build benchmark schema registry: %w", err)
-	}
-	metadata, err := schemameta.NewMetadataLoader(pool, schemaTable, FixturesDir()).LoadMetadata(ctx)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("load benchmark metadata: %w", err)
-	}
-
-	duckCfg := engineDuckDBConfig(h)
-	repo := internal.NewDBPersistentRecordRepository(pool, metadata, internal.WithPlanCache(r.planCache))
-	engine := fedengine.NewDBFederatedQueryEngine(
-		repo,
-		fedengine.NewPostgresDirtyIDFetcher(pool),
-		fedengine.NewDuckDBClientQueryExecutor(h.Duck),
-		nil,
-		duckCfg,
-		metadata,
-		fedengine.DuckDBPostgresConnStringFromPool(pool),
-		fedengine.WithPlanCache(r.planCache),
-	)
-
-	schemaName := workload.TargetSchema
-	if schemaName == "" {
-		schemaName = "trade"
-	}
-	schemaID, _, err := registry.GetSchemaAttributeCacheByName(schemaName)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("resolve schema %s: %w", schemaName, err)
-	}
-
-	pageSize := workload.PageSize
-	if pageSize <= 0 {
-		pageSize = r.config.PageSize
-	}
-	pageNumber := workload.PageNumber
-	if pageNumber < 1 {
-		pageNumber = 1
-	}
-
-	tables := model.StorageTables{
-		EAVData:    h.CDCConfig.EAVDataTable,
-		EntityMain: h.CDCConfig.EntityMainTable,
-		ChangeLog:  h.CDCConfig.ChangeLogTable,
-	}
-
-	// Build attribute orders for sort
-	attrOrders := []model.AttributeOrder{}
-	if cache, ok := metadata.GetSchemaCacheByID(schemaID); ok {
-		if meta, found := cache["tradeTime"]; found {
-			attrOrders = append(attrOrders, model.AttributeOrder{
-				AttrID:    meta.AttributeID,
-				ValueType: meta.ValueType,
-				SortOrder: forma.SortOrderDesc,
-			})
-		}
-	}
-	_ = attrOrders
-
-	// Build the federated query with filter conditions
-	fq := &model.FederatedAttributeQuery{
-		AttributeQuery: model.AttributeQuery{
-			SchemaID: schemaID,
-			Limit:    pageSize,
-			Offset:   0,
-		},
-	}
-	if conditions := workload.ResolvedFilterConditions(); len(conditions) > 0 {
-		fq.Condition = conditionForWorkload(workload)
-	}
-	if h.Duck != nil {
-		fq.DuckDBHints = &model.DuckDBRenderHints{
-			S3ParquetPathTemplate: benchmarkS3ParquetPathTemplate(h),
-		}
-	}
-
-	start := time.Now()
-	var totalRecords int64
-
-	if pageNumber == 1 {
-		// Page 1: keyset with no cursor is equivalent to offset 0.
-		// For page 1 of a keyset workload, use a nil cursor.
-		recs, total, err := engine.ExecuteDuckDBFederatedQuery(ctx, tables, fq, pageSize, 0, nil, &model.FederatedQueryOptions{
-			IncludeExecutionPlan: true,
-		})
-		if err != nil {
-			return nil, nil, nil, fmt.Errorf("keyset page 1: %w", err)
-		}
-		totalRecords = total
-		result := &forma.QueryResult{TotalRecords: int(totalRecords), ItemsPerPage: pageSize, CurrentPage: 1}
-		return result, recs, nil, nil
-	}
-
-	// For page N > 1: fetch the cursor row (page N-1's last row) via offset,
-	// then use keyset to fetch page N.
-	cursorOffset := (pageNumber-1)*pageSize - 1
-	cursorFq := *fq
-	cursorFq.Limit = 1
-	cursorFq.Offset = cursorOffset
-	cursorRecs, _, err := engine.ExecuteDuckDBFederatedQuery(ctx, tables, &cursorFq, 1, cursorOffset, nil, &model.FederatedQueryOptions{MaxRows: 1})
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("fetch cursor row: %w", err)
-	}
-	if len(cursorRecs) == 0 {
-		result := &forma.QueryResult{TotalRecords: 0, ItemsPerPage: pageSize, CurrentPage: pageNumber}
-		return result, nil, nil, nil
-	}
-
-	// Build the keyset cursor from the cursor row.
-	// Benchmark schema sorts by created_at (trade_time) DESC + row_id ASC tiebreaker.
-	cursor := &model.KeysetCursor{
-		Columns: []model.KeysetColumn{
-			{Attribute: "created_at", Direction: forma.SortOrderDesc},
-			{Attribute: "row_id", Direction: forma.SortOrderAsc},
-		},
-		Values: []interface{}{
-			cursorRecs[0].CreatedAt,
-			cursorRecs[0].RowID.String(),
-		},
-		Mode: model.KeysetCursorModeAfter,
-	}
-
-	fq.KeysetCursor = cursor
-	opts := &model.FederatedQueryOptions{
-		KeysetEnabled:        true,
-		IncludeExecutionPlan: true,
-	}
-
-	recs, total, err := engine.ExecuteFederatedPaginatedQuery(ctx, tables, fq, pageSize, 0, nil, opts)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("keyset query page %d: %w", pageNumber, err)
-	}
-	totalRecords = total
-
-	plan := opts.ExecutionPlan
-	if plan == nil {
-		plan = &model.ExecutionPlan{Notes: []string{"keyset_pagination"}, Timings: map[string]int64{}}
-	} else {
-		plan.Notes = append(plan.Notes, "keyset_pagination")
-	}
-	_ = start
-
-	result := &forma.QueryResult{
-		TotalRecords: int(totalRecords),
-		ItemsPerPage: pageSize,
-		CurrentPage:  pageNumber,
-	}
-	return result, recs, plan, nil
-}
-
-// engineDuckDBConfig returns the engine-level DuckDB config for a harness:
-// the exact configuration the harness started DuckDB with (single source of
-// truth for resource settings), or a zero config when DuckDB is not running.
-func engineDuckDBConfig(h *federated.FederatedTestHarness) forma.DuckDBConfig {
-	if h.Duck == nil {
-		return forma.DuckDBConfig{}
-	}
-	return h.DuckCfg
-}
-
-func executeServiceQuery(ctx context.Context, h *federated.FederatedTestHarness, req *forma.QueryRequest, defaultPageSize int, planCache *queryplan.Cache) (*forma.QueryResult, []*model.PersistentRecord, error) {
-	if h == nil || h.PGDSN == "" {
-		return nil, nil, fmt.Errorf("benchmark harness postgres DSN is required")
-	}
-	pool, err := pgxpool.New(ctx, h.PGDSN)
-	if err != nil {
-		return nil, nil, fmt.Errorf("connect benchmark pgx pool: %w", err)
-	}
-	defer pool.Close()
-
-	if err := RegisterFixtureSchemas(h); err != nil {
-		return nil, nil, fmt.Errorf("register fixture schemas: %w", err)
-	}
-	schemaTable, err := ensureBenchmarkSchemaRegistry(ctx, pool)
-	if err != nil {
-		return nil, nil, fmt.Errorf("prepare benchmark schema registry: %w", err)
-	}
-
-	registry, err := schemameta.NewFileSchemaRegistry(pool, schemaTable, FixturesDir())
-	if err != nil {
-		return nil, nil, fmt.Errorf("build benchmark schema registry: %w", err)
-	}
-	metadata, err := schemameta.NewMetadataLoader(pool, schemaTable, FixturesDir()).LoadMetadata(ctx)
-	if err != nil {
-		return nil, nil, fmt.Errorf("load benchmark metadata: %w", err)
-	}
-
-	duckCfg := engineDuckDBConfig(h)
-
-	repo := internal.NewDBPersistentRecordRepository(pool, metadata, internal.WithPlanCache(planCache))
-	engine := fedengine.NewDBFederatedQueryEngine(
-		repo,
-		fedengine.NewPostgresDirtyIDFetcher(pool),
-		fedengine.NewDuckDBClientQueryExecutor(h.Duck),
-		nil,
-		duckCfg,
-		metadata,
-		fedengine.DuckDBPostgresConnStringFromPool(pool),
-		fedengine.WithPlanCache(planCache),
-	)
-	config := &forma.Config{
-		Database: forma.DatabaseConfig{
-			TableNames: forma.TableNames{
-				SchemaRegistry: schemaTable,
-				EntityMain:     h.CDCConfig.EntityMainTable,
-				EAVData:        h.CDCConfig.EAVDataTable,
-				ChangeLog:      h.CDCConfig.ChangeLogTable,
-			},
-		},
-		Query: forma.QueryConfig{
-			DefaultPageSize: benchmarkDefaultPageSize(defaultPageSize),
-			MaxPageSize:     maxInt(defaultPageSize, 1000),
-		},
-		Entity: forma.EntityConfig{
-			SchemaDirectory: FixturesDir(),
-		},
-		DuckDB: duckCfg,
-	}
-	transformer := transform.NewPersistentRecordTransformer(registry)
-	manager := internal.NewEntityManager(transformer, repo, engine, registry, config)
-	if req != nil && req.Federated != nil && req.Federated.Enabled {
-		req.Federated.S3ParquetPathTemplate = benchmarkS3ParquetPathTemplate(h)
-	}
-	result, err := manager.Query(ctx, req)
-	if err != nil {
-		return nil, nil, fmt.Errorf("execute service query: %w", err)
-	}
-	records, err := persistentRecordsForQueryResult(ctx, result, registry)
-	if err != nil {
-		return nil, nil, err
-	}
-	return result, records, nil
-}
-
-func executeServiceQueryWithPlan(ctx context.Context, h *federated.FederatedTestHarness, req *forma.QueryRequest, defaultPageSize int, capturePlan bool, planCache *queryplan.Cache) (*forma.QueryResult, []*model.PersistentRecord, *model.ExecutionPlan, error) {
-	if h == nil || h.PGDSN == "" {
-		return nil, nil, nil, fmt.Errorf("benchmark harness postgres DSN is required")
-	}
-	pool, err := pgxpool.New(ctx, h.PGDSN)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("connect benchmark pgx pool: %w", err)
-	}
-	defer pool.Close()
-
-	result, records, err := executeServiceQuery(ctx, h, req, defaultPageSize, planCache)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	if !capturePlan || req == nil || req.Federated == nil {
-		return result, records, nil, nil
-	}
-
-	schemaTable, err := ensureBenchmarkSchemaRegistry(ctx, pool)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("prepare benchmark schema registry: %w", err)
-	}
-	registry, err := schemameta.NewFileSchemaRegistry(pool, schemaTable, FixturesDir())
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("build benchmark schema registry: %w", err)
-	}
-	metadata, err := schemameta.NewMetadataLoader(pool, schemaTable, FixturesDir()).LoadMetadata(ctx)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("load benchmark metadata: %w", err)
-	}
-
-	schemaName := req.SchemaName
-	if schemaName == "" {
-		schemaName = "trade"
-	}
-	schemaID, _, err := registry.GetSchemaAttributeCacheByName(schemaName)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("resolve schema %s: %w", schemaName, err)
-	}
-
-	duckCfg := engineDuckDBConfig(h)
-	repo := internal.NewDBPersistentRecordRepository(pool, metadata, internal.WithPlanCache(planCache))
-	engine := fedengine.NewDBFederatedQueryEngine(
-		repo,
-		fedengine.NewPostgresDirtyIDFetcher(pool),
-		fedengine.NewDuckDBClientQueryExecutor(h.Duck),
-		nil,
-		duckCfg,
-		metadata,
-		fedengine.DuckDBPostgresConnStringFromPool(pool),
-		fedengine.WithPlanCache(planCache),
-	)
-
-	pageSize := benchmarkDefaultPageSize(defaultPageSize)
-	limit := pageSize
-	offset := (maxInt(req.Page, 1) - 1) * pageSize
-
-	sortOrder := forma.SortOrderDesc
-	if req.SortOrder == forma.SortOrderAsc {
-		sortOrder = forma.SortOrderAsc
-	}
-
-	var attrOrders []model.AttributeOrder
-	if cache, ok := metadata.GetSchemaCacheByID(schemaID); ok {
-		for _, sortAttr := range req.SortBy {
-			meta, found := cache[sortAttr]
-			if !found {
-				continue
-			}
-			order := model.AttributeOrder{
-				AttrID:    meta.AttributeID,
-				ValueType: meta.ValueType,
-				SortOrder: sortOrder,
-			}
-			if meta.ColumnBinding != nil {
-				order.StorageLocation = forma.AttributeStorageLocationMain
-				order.ColumnName = string(meta.ColumnBinding.ColumnName)
-			} else {
-				order.StorageLocation = forma.AttributeStorageLocationEAV
-			}
-			attrOrders = append(attrOrders, order)
-		}
-	}
-
-	plan := &model.ExecutionPlan{Timings: map[string]int64{}, Notes: []string{}}
-	fqOpts := &model.FederatedQueryOptions{
-		IncludeExecutionPlan: true,
-		ExecutionPlan:        plan,
-	}
-
-	tables := model.StorageTables{
-		EntityMain: h.CDCConfig.EntityMainTable,
-		EAVData:    h.CDCConfig.EAVDataTable,
-		ChangeLog:  h.CDCConfig.ChangeLogTable,
-	}
-
-	fq := &model.FederatedAttributeQuery{
-		AttributeQuery: model.AttributeQuery{
-			SchemaID:        schemaID,
-			Condition:       req.Condition,
-			AttributeOrders: attrOrders,
-			Limit:           limit,
-			Offset:          offset,
-		},
-		PreferredTiers: []model.DataTier{model.DataTierHot, model.DataTierWarm, model.DataTierCold},
-	}
-	if req.Federated.S3ParquetPathTemplate != "" {
-		fq.DuckDBHints = &model.DuckDBRenderHints{S3ParquetPathTemplate: req.Federated.S3ParquetPathTemplate}
-	}
-	if req.Federated.PreferHot {
-		fq.PreferHot = true
-	}
-
-	_, err = engine.Query(ctx, tables, fq, fqOpts)
-	if err != nil {
-		return result, records, plan, nil
-	}
-	if fqOpts.ExecutionPlan == nil || !fqOpts.ExecutionPlan.Routing.UseDuckDB {
-		return result, records, plan, fmt.Errorf("federated benchmark query did not route to duckdb")
-	}
-
-	return result, records, plan, nil
-}
-
-func ensureBenchmarkSchemaRegistry(ctx context.Context, pool *pgxpool.Pool) (string, error) {
-	const tableName = "benchmark_schema_registry"
-	if _, err := pool.Exec(ctx, `
-		CREATE TABLE IF NOT EXISTS benchmark_schema_registry (
-			schema_id SMALLINT PRIMARY KEY,
-			schema_name TEXT NOT NULL UNIQUE,
-			created_at BIGINT NOT NULL DEFAULT EXTRACT(EPOCH FROM NOW()) * 1000
-		)`); err != nil {
-		return "", err
-	}
-	if _, err := pool.Exec(ctx, `DELETE FROM benchmark_schema_registry`); err != nil {
-		return "", err
-	}
-	for _, fixture := range DefaultSchemaFixtures() {
-		if _, err := pool.Exec(ctx, `INSERT INTO benchmark_schema_registry (schema_id, schema_name) VALUES ($1, $2)`, fixture.ID, fixture.Name); err != nil {
-			return "", err
-		}
-	}
-	return tableName, nil
-}
-
-func queryRequestForWorkload(workload WorkloadDefinition, defaultPageSize int) (*forma.QueryRequest, int) {
-	pageSize := workload.PageSize
-	if pageSize <= 0 {
-		pageSize = defaultPageSize
-	}
-	req := &forma.QueryRequest{
-		SchemaName:   workload.TargetSchema,
-		Page:         maxInt(workload.PageNumber, 1),
-		ItemsPerPage: pageSize,
-	}
-	if workload.ExecutionSource == "service" {
-		req.Federated = &forma.FederatedQueryRequest{
-			Enabled:              true,
-			PreferredTiers:       []string{"hot", "warm", "cold"},
-			IncludeExecutionPlan: workload.Category == WorkloadCategoryPushdown || workload.Category == WorkloadCategoryTierMix,
-		}
-	}
-	if workload.TargetSchema == "trade" {
-		req.SortBy = []string{"tradeTime"}
-		req.SortOrder = forma.SortOrderDesc
-	}
-	if cond := conditionForWorkload(workload); cond != nil {
-		req.Condition = cond
-	}
-	return req, pageSize
-}
-
-func benchmarkS3ParquetPathTemplate(h *federated.FederatedTestHarness) string {
-	if h == nil {
-		return ""
-	}
-	return fmt.Sprintf("s3://%s/%s/{{.SchemaID}}/base/*.parquet, s3://%s/%s/{{.SchemaID}}/delta/*.parquet", h.S3Bucket, h.S3Prefix, h.S3Bucket, h.S3Prefix)
-}
-
-func conditionForWorkload(workload WorkloadDefinition) forma.Condition {
-	conditions := make([]forma.Condition, 0, len(workload.ResolvedFilterConditions())+1)
-	for key, value := range workload.ResolvedFilterConditions() {
-		conditions = append(conditions, &forma.KvCondition{Attr: key, Value: fmt.Sprintf("equals:%v", value)})
-	}
-	if len(conditions) == 0 {
-		return nil
-	}
-	if len(conditions) == 1 {
-		return conditions[0]
-	}
-	return &forma.CompositeCondition{Logic: forma.LogicAnd, Conditions: conditions}
-}
-
-func persistentRecordsForQueryResult(ctx context.Context, result *forma.QueryResult, registry forma.SchemaRegistry) ([]*model.PersistentRecord, error) {
-	if result == nil {
-		return nil, nil
-	}
-	transformer := transform.NewPersistentRecordTransformer(registry)
-	records := make([]*model.PersistentRecord, 0, len(result.Data))
-	for _, data := range result.Data {
-		if data == nil {
-			continue
-		}
-		schemaID, _, err := registry.GetSchemaAttributeCacheByName(data.SchemaName)
-		if err != nil {
-			return nil, fmt.Errorf("resolve schema %s: %w", data.SchemaName, err)
-		}
-		record, err := transformer.ToPersistentRecord(ctx, schemaID, data.RowID, data.Attributes)
-		if err != nil {
-			return nil, fmt.Errorf("rebuild persistent record %s: %w", data.RowID, err)
-		}
-		if record != nil {
-			records = append(records, record)
-		}
-	}
-	return records, nil
-}
-
-func benchmarkDefaultPageSize(pageSize int) int {
-	if pageSize > 0 {
-		return pageSize
-	}
-	return 20
-}
-
-func failedWorkloadRunResult(workload WorkloadDefinition, distribution Distribution, defaultPageSize int, infraError string) WorkloadRunResult {
-	pageSize := workload.PageSize
-	if pageSize <= 0 {
-		pageSize = defaultPageSize
-	}
-	return WorkloadRunResult{
-		Name:         workload.Name,
-		Category:     string(workload.Category),
-		Distribution: distribution,
-		PageSize:     pageSize,
-		PageNumber:   workload.PageNumber,
-		Offset:       workload.DerivedOffset(defaultPageSize),
-		PreferHot:    workload.PreferHot,
-		Passed:       false,
-		FailureKind:  FailureKindInfra,
-		FailureCount: 1,
-		InfraError:   infraError,
-	}
-}
-
-type expectedWorkloadResult struct {
-	TotalRecords int64
-	RowIDs       []string
-}
-
-type workloadSemantics struct {
-	TradeTimeStart int64
-	TradeTimeEnd   int64
-}
-
-func buildExpectedWorkloadResults(dataset *GeneratedDataset, workloads []WorkloadDefinition, defaultPageSize int) map[string]expectedWorkloadResult {
-	if dataset == nil {
-		return map[string]expectedWorkloadResult{}
-	}
-	return buildExpectedWorkloadResultsFromRecords(dataset.Records, workloads, defaultPageSize, dataset.Config, nil)
-}
-
-// workloadExecutesPostgresOnly mirrors the execution override in executeWorkload:
-// prefer-hot tier-mix workloads run against the Postgres hot buffer only, so
-// their loaded-state oracle must be restricted to hot-tier records.
-func workloadExecutesPostgresOnly(workload WorkloadDefinition) bool {
-	return workload.PreferHot && workload.Category == WorkloadCategoryTierMix
-}
-
-func buildExpectedWorkloadResultsFromRecords(records []GeneratedRecord, workloads []WorkloadDefinition, defaultPageSize int, genCfg GeneratorConfig, hotKeys map[string]struct{}) map[string]expectedWorkloadResult {
-	results := make(map[string]expectedWorkloadResult, len(workloads))
-	visible := expectedVisibleRecords(records)
-	var hotVisible []GeneratedRecord
-	if hotKeys != nil {
-		hotVisible = make([]GeneratedRecord, 0, len(hotKeys))
-		for _, record := range visible {
-			if _, ok := hotKeys[schemaRowKey(record.SchemaID, record.RowID)]; ok {
-				hotVisible = append(hotVisible, record)
-			}
-		}
-	}
-	for _, workload := range workloads {
-		source := visible
-		if hotKeys != nil && workloadExecutesPostgresOnly(workload) {
-			source = hotVisible
-		}
-		matching := filterExpectedRecordsForWorkload(source, workload, semanticsForWorkload(workload, genCfg))
-		sortExpectedRecordsForWorkload(matching, workload)
-		pageSize := workload.PageSize
-		if pageSize <= 0 {
-			pageSize = defaultPageSize
-		}
-		offset := workload.DerivedOffset(defaultPageSize)
-		rowIDs := expectedPageRowIDs(matching, offset, pageSize)
-		results[workload.Name] = expectedWorkloadResult{TotalRecords: int64(len(matching)), RowIDs: rowIDs}
-	}
-	return results
-}
-
-func buildLoadedStateSnapshot(ctx context.Context, h *federated.FederatedTestHarness, dataset *TieredDataset) ([]GeneratedRecord, map[string]struct{}, error) {
-	if h == nil || dataset == nil {
-		return nil, nil, fmt.Errorf("harness and dataset are required")
-	}
-	hotRecords, hotKeys, err := loadHotStateRecords(ctx, h)
-	if err != nil {
-		return nil, nil, err
-	}
-	records := make([]GeneratedRecord, 0, len(dataset.Base)+len(dataset.Delta)+len(hotRecords))
-	for _, bucket := range [][]GeneratedRecord{dataset.Base, dataset.Delta} {
-		for _, record := range bucket {
-			if _, ok := hotKeys[schemaRowKey(record.SchemaID, record.RowID)]; ok {
-				continue
-			}
-			records = append(records, cloneGeneratedRecord(record))
-		}
-	}
-	records = append(records, hotRecords...)
-	return records, hotKeys, nil
-}
-
-func loadHotStateRecords(ctx context.Context, h *federated.FederatedTestHarness) ([]GeneratedRecord, map[string]struct{}, error) {
-	registry, err := LoadFixtureRegistry()
-	if err != nil {
-		return nil, nil, fmt.Errorf("load fixture registry: %w", err)
-	}
-	_, tradeCache, err := registry.GetSchemaAttributeCacheByID(SchemaIDTrade)
-	if err != nil {
-		return nil, nil, fmt.Errorf("load trade schema cache: %w", err)
-	}
-	tradeAttrID := func(name string) (int16, error) {
-		meta, ok := tradeCache[name]
-		if !ok {
-			return 0, fmt.Errorf("trade attribute %s not found in fixture cache", name)
-		}
-		return meta.AttributeID, nil
-	}
-	symbolAttrID, err := tradeAttrID("symbol")
-	if err != nil {
-		return nil, nil, err
-	}
-	exchangeAttrID, err := tradeAttrID("exchange")
-	if err != nil {
-		return nil, nil, err
-	}
-	regionAttrID, err := tradeAttrID("region")
-	if err != nil {
-		return nil, nil, err
-	}
-	tradeTypeAttrID, err := tradeAttrID("tradeType")
-	if err != nil {
-		return nil, nil, err
-	}
-	tradeTimeAttrID, err := tradeAttrID("tradeTime")
-	if err != nil {
-		return nil, nil, err
-	}
-	rows, err := h.PGDB.QueryContext(ctx, `
-		SELECT cl.schema_id, cl.row_id, cl.changed_at, COALESCE(cl.deleted_at, 0),
-			em.text_01, em.text_02, em.smallint_01, em.bigint_02,
-			hot_vals.symbol, hot_vals.exchange, hot_vals.region, hot_vals.trade_type, hot_vals.trade_time, hot_vals.name
-		FROM change_log cl
-		LEFT JOIN entity_main em
-			ON em.ltbase_schema_id = cl.schema_id AND em.ltbase_row_id = cl.row_id
-		LEFT JOIN (
-			SELECT schema_id, row_id,
-				MAX(CASE WHEN attr_id = $1 THEN value_text END) AS symbol,
-				MAX(CASE WHEN attr_id = $2 THEN value_text END) AS exchange,
-				MAX(CASE WHEN attr_id = $3 THEN value_text END) AS region,
-				MAX(CASE WHEN attr_id = $4 THEN value_numeric END) AS trade_type,
-				MAX(CASE WHEN attr_id = $5 THEN COALESCE(value_text, CAST(value_numeric AS TEXT)) END) AS trade_time,
-				MAX(CASE WHEN attr_id = $6 THEN value_text END) AS name
-			FROM eav_data
-			GROUP BY schema_id, row_id
-		) hot_vals ON hot_vals.schema_id = cl.schema_id AND hot_vals.row_id = cl.row_id
-		WHERE cl.flushed_at = 0
-	`,
-		symbolAttrID,
-		exchangeAttrID,
-		regionAttrID,
-		tradeTypeAttrID,
-		tradeTimeAttrID,
-		0,
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("load hot state snapshot: %w", err)
-	}
-	defer rows.Close()
-	records := make([]GeneratedRecord, 0)
-	keys := make(map[string]struct{})
-	for rows.Next() {
-		record, err := scanLoadedHotRecord(rows)
-		if err != nil {
-			return nil, nil, err
-		}
-		records = append(records, record)
-		keys[schemaRowKey(record.SchemaID, record.RowID)] = struct{}{}
-	}
-	if err := rows.Err(); err != nil {
-		return nil, nil, fmt.Errorf("iterate hot state snapshot: %w", err)
-	}
-	return records, keys, nil
-}
-
-func scanLoadedHotRecord(rows *sql.Rows) (GeneratedRecord, error) {
-	var schemaID int16
-	var rowID uuid.UUID
-	var changedAt int64
-	var deletedAt int64
-	var text01 sql.NullString
-	var text02 sql.NullString
-	var smallint01 sql.NullInt16
-	var bigint02 sql.NullInt64
-	var symbol sql.NullString
-	var exchange sql.NullString
-	var region sql.NullString
-	var tradeType sql.NullFloat64
-	var tradeTime any
-	var name sql.NullString
-	if err := rows.Scan(&schemaID, &rowID, &changedAt, &deletedAt, &text01, &text02, &smallint01, &bigint02, &symbol, &exchange, &region, &tradeType, &tradeTime, &name); err != nil {
-		return GeneratedRecord{}, fmt.Errorf("scan hot state row: %w", err)
-	}
-	attrs := make(map[string]any)
-	schemaName, err := schemaNameForID(schemaID)
-	if err != nil {
-		return GeneratedRecord{}, err
-	}
-	switch schemaID {
-	case SchemaIDTrade:
-		if symbol.Valid {
-			attrs["symbol"] = symbol.String
-		} else if text01.Valid {
-			attrs["symbol"] = text01.String
-		}
-		if exchange.Valid {
-			attrs["exchange"] = exchange.String
-		}
-		if region.Valid {
-			attrs["region"] = region.String
-		} else if text02.Valid {
-			attrs["region"] = text02.String
-		}
-		if tradeType.Valid {
-			attrs["tradeType"] = int64(tradeType.Float64)
-		} else if smallint01.Valid {
-			attrs["tradeType"] = int64(smallint01.Int16)
-		}
-		if normalizedTradeTime := normalizeBenchmarkTradeTime(tradeTime); normalizedTradeTime != "" {
-			attrs["tradeTime"] = normalizedTradeTime
-		} else if bigint02.Valid {
-			attrs["tradeTime"] = strconv.FormatInt(bigint02.Int64, 10)
-		}
-		if name.Valid {
-			attrs["name"] = name.String
-		} else if symbol.Valid {
-			attrs["name"] = symbol.String
-		}
-	case SchemaIDCustomer:
-		if text02.Valid {
-			attrs["region"] = text02.String
-		}
-		if name.Valid {
-			attrs["name"] = name.String
-		} else if text01.Valid {
-			attrs["name"] = text01.String
-		}
-	case SchemaIDSecurity:
-		if symbol.Valid {
-			attrs["symbol"] = symbol.String
-		} else if text01.Valid {
-			attrs["symbol"] = text01.String
-		}
-		if name.Valid {
-			attrs["companyName"] = name.String
-		}
-	}
-	return GeneratedRecord{SchemaID: schemaID, SchemaName: schemaName, RowID: rowID, Version: 0, ChangedAt: changedAt, DeletedAt: deletedAt, Attributes: attrs}, nil
-}
-
-func normalizeBenchmarkTradeTime(value any) string {
-	switch v := value.(type) {
-	case time.Time:
-		return strconv.FormatInt(v.UnixMilli(), 10)
-	case sql.NullTime:
-		if v.Valid {
-			return strconv.FormatInt(v.Time.UnixMilli(), 10)
-		}
-	case string:
-		if unixMillis, err := strconv.ParseInt(v, 10, 64); err == nil {
-			return strconv.FormatInt(unixMillis, 10)
-		}
-		if parsed, err := time.Parse(time.RFC3339, v); err == nil {
-			return strconv.FormatInt(parsed.UnixMilli(), 10)
-		}
-	case []byte:
-		return normalizeBenchmarkTradeTime(string(v))
-	}
-	return ""
-}
-
-func schemaNameForID(schemaID int16) (string, error) {
-	for _, fixture := range DefaultSchemaFixtures() {
-		if fixture.ID == schemaID {
-			return fixture.Name, nil
-		}
-	}
-	return "", fmt.Errorf("unknown benchmark schema id %d", schemaID)
-}
-
-// truthPassSampleStats reports how truth-pass verification was bounded.
-type truthPassSampleStats struct {
-	Applied    bool
-	Cap        int
-	Candidates int
-	Sampled    int
-}
-
-// truthPassCollectPageSize bounds each batched truth-pass sweep page. It is far
-// larger than any small/medium-live hit set, so selective workloads finish in a
-// single round trip.
-const truthPassCollectPageSize = 1000
-
-// collectVisibleRowIDs pulls the full set of row IDs the engine returns for a
-// workload's filter conditions in one paginated sweep (no per-row RowID
-// filter). A candidate is visible iff its row id appears here, which is exactly
-// the membership the retired per-candidate Limit:1 + RowID probe tested. h.SchemaID
-// must already be set to the workload's schema by the caller.
-func collectVisibleRowIDs(ctx context.Context, h *federated.FederatedTestHarness, workload WorkloadDefinition) (map[uuid.UUID]struct{}, error) {
-	visible := make(map[uuid.UUID]struct{})
-	conditions := workload.ResolvedFilterConditions()
-	for offset := 0; ; offset += truthPassCollectPageSize {
-		result, err := h.ExecuteFederatedQuery(ctx, &federated.QueryOptions{
-			Limit:    truthPassCollectPageSize,
-			Offset:   offset,
-			Filter:   &federated.Filter{Conditions: conditions},
-			SortBy:   "tradeTime",
-			SortDesc: true,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("execute federated truth query for workload %s (offset %d): %w", workload.Name, offset, err)
-		}
-		for _, rec := range result.Records {
-			if rec != nil {
-				visible[rec.RowID] = struct{}{}
-			}
-		}
-		if len(result.Records) < truthPassCollectPageSize || offset+len(result.Records) >= int(result.TotalRecords) {
-			break
-		}
-	}
-	return visible, nil
-}
-
-// perCandidateVisible probes one candidate's visibility through the engine with
-// a RowID-filtered query. Capped (sampled) truth-pass mode uses this so its cost
-// stays bounded by the sample cap; the uncapped path uses collectVisibleRowIDs.
-func perCandidateVisible(ctx context.Context, h *federated.FederatedTestHarness, workload WorkloadDefinition, candidate GeneratedRecord) (bool, error) {
-	result, err := h.ExecuteFederatedQuery(ctx, &federated.QueryOptions{
-		Limit: 1,
-		Filter: &federated.Filter{
-			RowID:      candidate.RowID,
-			Conditions: workload.ResolvedFilterConditions(),
-		},
-		SortBy:   "tradeTime",
-		SortDesc: true,
-	})
-	if err != nil {
-		return false, fmt.Errorf("execute federated truth query for candidate %s: %w", candidate.RowID, err)
-	}
-	return result.TotalRecords > 0, nil
-}
-
-func buildExpectedWorkloadResultFromFederatedTruth(ctx context.Context, h *federated.FederatedTestHarness, workload WorkloadDefinition, defaultPageSize int, loadedRecords []GeneratedRecord, genCfg GeneratorConfig, sampleCap int, seed int64) (expectedWorkloadResult, truthPassSampleStats, error) {
-	if h == nil {
-		return expectedWorkloadResult{}, truthPassSampleStats{}, fmt.Errorf("harness cannot be nil")
-	}
-	semantics := semanticsForWorkload(workload, genCfg)
-	candidates := filterExpectedRecordsForWorkload(expectedVisibleRecords(loadedRecords), workload, semantics)
-	sortExpectedRecordsForWorkload(candidates, workload)
-	previousSchemaID := h.SchemaID
-	schemaID, err := workloadSchemaID(workload.TargetSchema)
-	if err != nil {
-		return expectedWorkloadResult{}, truthPassSampleStats{}, fmt.Errorf("resolve schema id for workload %s: %w", workload.Name, err)
-	}
-	h.SchemaID = schemaID
-	defer func() {
-		h.SchemaID = previousSchemaID
-	}()
-	// Uncapped mode verifies every candidate, so replace O(candidates)
-	// per-candidate RowID probes with one batched visibility sweep: membership
-	// is identical (a row is in the swept set iff a RowID-filtered query for it
-	// under the same conditions would return a row). Capped (heavy-live) mode
-	// only spot-checks a bounded sample, so it keeps the per-candidate probe —
-	// sweeping the full hit set there would materialize O(hits) row ids and
-	// break the sample cap's cost/memory bound.
-	var isVisible func(context.Context, GeneratedRecord) (bool, error)
-	if truthPassCapped(len(candidates), sampleCap) {
-		isVisible = func(ctx context.Context, candidate GeneratedRecord) (bool, error) {
-			return perCandidateVisible(ctx, h, workload, candidate)
-		}
-	} else {
-		visible, err := collectVisibleRowIDs(ctx, h, workload)
-		if err != nil {
-			return expectedWorkloadResult{}, truthPassSampleStats{}, err
-		}
-		isVisible = func(_ context.Context, candidate GeneratedRecord) (bool, error) {
-			_, ok := visible[candidate.RowID]
-			return ok, nil
-		}
-	}
-	return buildTruthPassExpected(ctx, isVisible, workload, defaultPageSize, candidates, sampleCap, seed)
-}
-
-// buildTruthPassExpected derives the expected result for a truth-pass
-// workload. Uncapped (sampleCap <= 0 or candidates <= cap) it verifies every
-// candidate and keeps only the visible ones — existing behavior. Capped, it
-// keeps the full reconstructed candidate set as the expected result and
-// verifies a seeded deterministic sample; a sampled candidate the engine
-// cannot see means reconstruction and engine truth diverge, which no
-// sampling rate can absorb, so the run fails hard.
-func buildTruthPassExpected(ctx context.Context, isVisible func(context.Context, GeneratedRecord) (bool, error), workload WorkloadDefinition, defaultPageSize int, candidates []GeneratedRecord, sampleCap int, seed int64) (expectedWorkloadResult, truthPassSampleStats, error) {
-	pageSize := workload.PageSize
-	if pageSize <= 0 {
-		pageSize = defaultPageSize
-	}
-	sampledIdx := selectTruthPassSampleIndices(seed, workload.Name, len(candidates), sampleCap)
-	if sampledIdx == nil {
-		matching := make([]GeneratedRecord, 0, len(candidates))
-		for _, candidate := range candidates {
-			visible, err := isVisible(ctx, candidate)
-			if err != nil {
-				return expectedWorkloadResult{}, truthPassSampleStats{}, err
-			}
-			if visible {
-				matching = append(matching, candidate)
-			}
-		}
-		rowIDs := expectedPageRowIDs(matching, workload.DerivedOffset(defaultPageSize), pageSize)
-		return expectedWorkloadResult{TotalRecords: int64(len(matching)), RowIDs: rowIDs}, truthPassSampleStats{Candidates: len(candidates), Sampled: len(candidates)}, nil
-	}
-	stats := truthPassSampleStats{Applied: true, Cap: sampleCap, Candidates: len(candidates), Sampled: len(sampledIdx)}
-	for i, candidate := range candidates {
-		if _, ok := sampledIdx[i]; !ok {
-			continue
-		}
-		visible, err := isVisible(ctx, candidate)
-		if err != nil {
-			return expectedWorkloadResult{}, stats, err
-		}
-		if !visible {
-			return expectedWorkloadResult{}, stats, fmt.Errorf("truth-pass spot check failed for workload %s: sampled candidate row_id=%s is not visible through the engine; reconstruction diverges from engine truth — investigate at a smaller scale without the sample cap before trusting sampled oracles", workload.Name, candidate.RowID)
-		}
-	}
-	rowIDs := expectedPageRowIDs(candidates, workload.DerivedOffset(defaultPageSize), pageSize)
-	return expectedWorkloadResult{TotalRecords: int64(len(candidates)), RowIDs: rowIDs}, stats, nil
-}
-
-func expectedVisibleRecords(records []GeneratedRecord) []GeneratedRecord {
-	latest := make(map[string]GeneratedRecord)
-	for _, record := range records {
-		key := schemaRowKey(record.SchemaID, record.RowID)
-		existing, ok := latest[key]
-		if ok && !recordWinsExpectedRecord(record, existing) {
-			continue
-		}
-		latest[key] = cloneGeneratedRecord(record)
-	}
-	out := make([]GeneratedRecord, 0, len(latest))
-	for _, record := range latest {
-		if record.DeletedAt > 0 {
-			continue
-		}
-		out = append(out, cloneGeneratedRecord(record))
-	}
-	return out
-}
-
-func recordWinsExpectedRecord(candidate, current GeneratedRecord) bool {
-	if candidate.ChangedAt != current.ChangedAt {
-		return candidate.ChangedAt > current.ChangedAt
-	}
-	if candidate.Version != current.Version {
-		return candidate.Version > current.Version
-	}
-	if candidate.DeletedAt != current.DeletedAt {
-		return candidate.DeletedAt > current.DeletedAt
-	}
-	return false
-}
-
-func filterExpectedRecordsForWorkload(records []GeneratedRecord, workload WorkloadDefinition, semantics workloadSemantics) []GeneratedRecord {
-	out := make([]GeneratedRecord, 0)
-	for _, record := range records {
-		if record.SchemaName != workload.TargetSchema {
-			continue
-		}
-		if semantics.TradeTimeStart > 0 || semantics.TradeTimeEnd > 0 {
-			tradeTime := generatedRecordTradeTime(record)
-			if semantics.TradeTimeStart > 0 && tradeTime < semantics.TradeTimeStart {
-				continue
-			}
-			if semantics.TradeTimeEnd > 0 && tradeTime > semantics.TradeTimeEnd {
-				continue
-			}
-		}
-		if len(workload.ResolvedFilterConditions()) > 0 && !generatedRecordMatchesFilterForWorkload(record, workload) {
-			continue
-		}
-		out = append(out, cloneGeneratedRecord(record))
-	}
-	return out
-}
-
-func semanticsForWorkload(workload WorkloadDefinition, genCfg GeneratorConfig) workloadSemantics {
-	if workload.TargetSchema != "trade" {
-		return workloadSemantics{}
-	}
-	baseTime := genCfg.BaseTime
-	if baseTime.IsZero() {
-		baseTime = defaultBaseTime
-	}
-	windowDays := genCfg.TimeWindowDays
-	if windowDays <= 0 {
-		windowDays = DefaultGeneratorConfig().TimeWindowDays
-	}
-	windowMillis := int64(windowDays) * 24 * int64(time.Hour/time.Millisecond)
-	var start, end int64
-	switch workload.Name {
-	case "mixed-tier-window":
-		start = baseTime.UnixMilli() - (windowMillis * 4 / 5)
-		end = baseTime.UnixMilli() - windowMillis/5
-	case "hot-only-window":
-		start = baseTime.UnixMilli() - windowMillis/5
-		end = baseTime.UnixMilli()
-	case "cold-only-window":
-		start = baseTime.UnixMilli() - windowMillis
-		end = baseTime.UnixMilli() - (windowMillis * 4 / 5)
-	default:
-		return workloadSemantics{}
-	}
-	if start > end {
-		start, end = end, start
-	}
-	return workloadSemantics{TradeTimeStart: start, TradeTimeEnd: end}
-}
-func generatedRecordMatchesFilterForWorkload(record GeneratedRecord, workload WorkloadDefinition) bool {
-	for key, expected := range workload.ResolvedFilterConditions() {
-		value, ok := benchmarkVisibleAttributeValue(record, key)
-		if !ok {
-			return false
-		}
-		if key == "tradeType" {
-			if fmt.Sprintf("%v", value) != fmt.Sprintf("%v", expected) {
-				return false
-			}
-			continue
-		}
-		if fmt.Sprint(value) != fmt.Sprint(expected) {
-			return false
-		}
-	}
-	return true
-}
-
-func benchmarkVisibleAttributeValue(record GeneratedRecord, attribute string) (any, bool) {
-	if value, ok := record.Attributes[attribute]; ok {
-		return value, true
-	}
-	if record.SchemaName != "trade" {
-		return nil, false
-	}
-	switch attribute {
-	case "symbol", "name":
-		value, ok := record.Attributes["symbol"]
-		return value, ok
-	case "tradeType":
-		value, ok := record.Attributes["tradeType"]
-		return value, ok
-	case "tradeTime":
-		value, ok := record.Attributes["tradeTime"]
-		return value, ok
-	default:
-		return nil, false
-	}
-}
-
-func sortExpectedRecordsForWorkload(records []GeneratedRecord, workload WorkloadDefinition) {
-	if workload.TargetSchema == "trade" {
-		sort.Slice(records, func(i, j int) bool {
-			left := generatedRecordTradeTime(records[i])
-			right := generatedRecordTradeTime(records[j])
-			if left != right {
-				return left > right
-			}
-			return records[i].RowID.String() < records[j].RowID.String()
-		})
-		return
-	}
-	sort.Slice(records, func(i, j int) bool {
-		return records[i].RowID.String() < records[j].RowID.String()
-	})
-}
-
-func generatedRecordTradeTime(record GeneratedRecord) int64 {
-	value, ok := record.Attributes["tradeTime"]
-	if !ok {
-		return 0
-	}
-	switch v := value.(type) {
-	case int64:
-		return v
-	case int:
-		return int64(v)
-	case string:
-		parsed, err := time.Parse(time.RFC3339, v)
-		if err == nil {
-			return parsed.UnixMilli()
-		}
-		if unixMillis, convErr := strconv.ParseInt(v, 10, 64); convErr == nil {
-			return unixMillis
-		}
-	}
-	return 0
-}
-
-func expectedPageRowIDs(records []GeneratedRecord, offset, pageSize int) []string {
-	if offset >= len(records) {
-		return nil
-	}
-	end := offset + pageSize
-	if end > len(records) {
-		end = len(records)
-	}
-	rowIDs := make([]string, 0, end-offset)
-	for _, record := range records[offset:end] {
-		rowIDs = append(rowIDs, record.RowID.String())
-	}
-	return rowIDs
-}
-
-func validateExpectedWorkloadOutcome(run WorkloadRunResult, expected expectedWorkloadResult) []AssertionResult {
-	assertions := []AssertionResult{{
-		Name:    "total-records-match-expected",
-		Passed:  run.TotalRecords == expected.TotalRecords,
-		Message: fmt.Sprintf("actual=%d expected=%d", run.TotalRecords, expected.TotalRecords),
-	}}
-	actualRows := append([]string(nil), run.RowIDs...)
-	expectedRows := append([]string(nil), expected.RowIDs...)
-	assertions = append(assertions, AssertionResult{
-		Name:    "page-row-ids-match-expected",
-		Passed:  stringSlicesEqual(actualRows, expectedRows),
-		Message: fmt.Sprintf("actual=%v expected=%v", actualRows, expectedRows),
-	})
-	return assertions
-}
-
-func stringSlicesEqual(left, right []string) bool {
-	if len(left) != len(right) {
-		return false
-	}
-	for i := range left {
-		if left[i] != right[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func failureKindForRun(run WorkloadRunResult) string {
-	if run.InfraError != "" {
-		return FailureKindInfra
-	}
-	if countFailedAssertions(run.Assertions) > 0 {
-		return FailureKindCorrectness
-	}
-	return ""
-}
-
-func validateBasicWorkloadAssertions(workload WorkloadDefinition, run WorkloadRunResult) []AssertionResult {
-	assertions := []AssertionResult{
-		{
-			Name:    "non-negative-total-records",
-			Passed:  run.TotalRecords >= 0,
-			Message: fmt.Sprintf("total_records=%d", run.TotalRecords),
-		},
-		{
-			Name:    "page-size-bound",
-			Passed:  run.ResultCount <= run.PageSize,
-			Message: fmt.Sprintf("result_count=%d page_size=%d", run.ResultCount, run.PageSize),
-		},
-		{
-			Name:    "result-count-within-total-records",
-			Passed:  int64(run.ResultCount) <= run.TotalRecords,
-			Message: fmt.Sprintf("result_count=%d total_records=%d", run.ResultCount, run.TotalRecords),
-		},
-		{
-			Name:    "empty-page-only-when-offset-reaches-total",
-			Passed:  run.ResultCount > 0 || run.Offset >= int(run.TotalRecords) || run.TotalRecords == 0,
-			Message: fmt.Sprintf("offset=%d total=%d result_count=%d", run.Offset, run.TotalRecords, run.ResultCount),
-		},
-	}
-	if workload.Category == WorkloadCategoryDeepPage {
-		assertions = append(assertions, AssertionResult{
-			Name:    "deep-page-empty-when-offset-exceeds-total",
-			Passed:  run.Offset < int(run.TotalRecords) || run.ResultCount == 0,
-			Message: fmt.Sprintf("offset=%d total=%d result_count=%d", run.Offset, run.TotalRecords, run.ResultCount),
-		})
-	}
-	return assertions
-}
-
-func validatePaginationTransition(previous, current WorkloadRunResult) []AssertionResult {
-	assertions := []AssertionResult{{
-		Name:    "non-decreasing-offsets-across-pagination-runs",
-		Passed:  current.Offset >= previous.Offset,
-		Message: fmt.Sprintf("previous_offset=%d current_offset=%d", previous.Offset, current.Offset),
-	}}
-	if current.Offset > previous.Offset {
-		overlap := hasRowIDOverlap(previous.RowIDs, current.RowIDs)
-		assertions = append(assertions, AssertionResult{
-			Name:    "no-overlap-across-page-slices",
-			Passed:  !overlap,
-			Message: fmt.Sprintf("previous_rows=%d current_rows=%d", len(previous.RowIDs), len(current.RowIDs)),
-		})
-	}
-	return assertions
-}
-
-func validateRepeatedRunStability(previous, current WorkloadRunResult) []AssertionResult {
-	assertions := []AssertionResult{{
-		Name:    "repeated-run-failure-kind-stable",
-		Passed:  previous.FailureKind == current.FailureKind,
-		Message: fmt.Sprintf("previous=%s current=%s", previous.FailureKind, current.FailureKind),
-	}}
-	assertions = append(assertions, AssertionResult{
-		Name:    "repeated-run-total-records-stable",
-		Passed:  previous.TotalRecords == current.TotalRecords,
-		Message: fmt.Sprintf("previous=%d current=%d", previous.TotalRecords, current.TotalRecords),
-	})
-	assertions = append(assertions, AssertionResult{
-		Name:    "repeated-run-page-row-ids-stable",
-		Passed:  stringSlicesEqual(previous.RowIDs, current.RowIDs),
-		Message: fmt.Sprintf("previous=%v current=%v", previous.RowIDs, current.RowIDs),
-	})
-	return assertions
-}
-
-func validateConcurrentRunStability(runs []WorkloadRunResult) []AssertionResult {
-	if len(runs) < 2 {
-		return nil
-	}
-	reference := runs[0]
-	var assertions []AssertionResult
-	allPassed := true
-	for i := 1; i < len(runs); i++ {
-		allPassed = allPassed && reference.TotalRecords == runs[i].TotalRecords
-	}
-	assertions = append(assertions, AssertionResult{
-		Name:    "concurrent-run-total-records-stable",
-		Passed:  allPassed,
-		Message: fmt.Sprintf("reference=%d workers=%d", reference.TotalRecords, len(runs)),
-	})
-
-	allPassed = true
-	for i := 1; i < len(runs); i++ {
-		allPassed = allPassed && stringSlicesEqual(reference.RowIDs, runs[i].RowIDs)
-	}
-	assertions = append(assertions, AssertionResult{
-		Name:    "concurrent-run-page-row-ids-stable",
-		Passed:  allPassed,
-		Message: fmt.Sprintf("reference_rows=%d workers=%d", len(reference.RowIDs), len(runs)),
-	})
-
-	allPassed = true
-	for i := 1; i < len(runs); i++ {
-		allPassed = allPassed && reference.FailureKind == runs[i].FailureKind
-	}
-	assertions = append(assertions, AssertionResult{
-		Name:    "concurrent-run-failure-kind-stable",
-		Passed:  allPassed,
-		Message: fmt.Sprintf("reference=%s workers=%d", reference.FailureKind, len(runs)),
-	})
-
-	return assertions
-}
-
-func validateConcurrentRouteStability(runs []WorkloadRunResult) []AssertionResult {
-	if len(runs) < 2 {
-		return nil
-	}
-	reference := runs[0]
-	if reference.RouteEngine == "" {
-		return nil
-	}
-	divergentEngines := make(map[string]int)
-	for _, run := range runs {
-		if run.RouteEngine != "" {
-			divergentEngines[run.RouteEngine]++
-		}
-	}
-	stable := len(divergentEngines) <= 1
-	parts := make([]string, 0, len(divergentEngines))
-	for engine, count := range divergentEngines {
-		parts = append(parts, fmt.Sprintf("%s=%d", engine, count))
-	}
-	assertions := []AssertionResult{{
-		Name:    "concurrent-run-route-engine-stable",
-		Passed:  stable,
-		Message: fmt.Sprintf("engines: %s", partsToStr(parts)),
-	}}
-	return assertions
-}
-
-func partsToStr(parts []string) string {
-	if len(parts) == 0 {
-		return ""
-	}
-	result := parts[0]
-	for i := 1; i < len(parts); i++ {
-		result += ", " + parts[i]
-	}
-	return result
-}
-
-func validateResultLevelAssertions(workload WorkloadDefinition, run WorkloadRunResult, records []*model.PersistentRecord, semantics workloadSemantics) []AssertionResult {
-	assertions := []AssertionResult{validateUniqueRows(run), validateSchemaScope(workload, records)}
-	if len(workload.ResolvedFilterConditions()) > 0 {
-		assertions = append(assertions, validateFilterMatch(workload, records))
-	}
-	if semantics.TradeTimeStart > 0 || semantics.TradeTimeEnd > 0 {
-		assertions = append(assertions, validateTradeTimeWindow(records, semantics))
-	}
-	if workload.TargetSchema == "trade" && len(records) > 1 {
-		assertions = append(assertions, validateSortOrder(records, "tradeTime", true))
-	}
-	return assertions
-}
-
-func validateSchemaScope(workload WorkloadDefinition, records []*model.PersistentRecord) AssertionResult {
-	expectedSchemaID, err := workloadSchemaID(workload.TargetSchema)
-	if err != nil {
-		return AssertionResult{Name: "schema-scoped-results-match-target", Passed: false, Message: err.Error()}
-	}
-	for _, record := range records {
-		if record.SchemaID != expectedSchemaID {
-			return AssertionResult{Name: "schema-scoped-results-match-target", Passed: false, Message: fmt.Sprintf("expected_schema=%d actual_schema=%d row=%s", expectedSchemaID, record.SchemaID, record.RowID)}
-		}
-	}
-	return AssertionResult{Name: "schema-scoped-results-match-target", Passed: true, Message: fmt.Sprintf("schema=%s rows=%d", workload.TargetSchema, len(records))}
-}
-
-func queryOptionsForWorkload(workload WorkloadDefinition, defaultPageSize int) *federated.QueryOptions {
-	return queryOptionsForWorkloadWithConfig(workload, defaultPageSize, DefaultGeneratorConfig())
-}
-
-func queryOptionsForWorkloadWithConfig(workload WorkloadDefinition, defaultPageSize int, genCfg GeneratorConfig) *federated.QueryOptions {
-	pageSize := workload.PageSize
-	if pageSize <= 0 {
-		pageSize = defaultPageSize
-	}
-	opts := &federated.QueryOptions{Limit: pageSize, Offset: workload.DerivedOffset(defaultPageSize)}
-	if workload.TargetSchema == "trade" {
-		opts.SortBy = "tradeTime"
-		opts.SortDesc = true
-	}
-	semantics := semanticsForWorkload(workload, genCfg)
-	opts.TradeTimeStart = semantics.TradeTimeStart
-	opts.TradeTimeEnd = semantics.TradeTimeEnd
-	return opts
-}
-
-func validateUniqueRows(run WorkloadRunResult) AssertionResult {
-	seen := make(map[string]struct{}, len(run.RowIDs))
-	for _, rowID := range run.RowIDs {
-		if _, ok := seen[rowID]; ok {
-			return AssertionResult{Name: "unique-row-ids-within-page", Passed: false, Message: rowID}
-		}
-		seen[rowID] = struct{}{}
-	}
-	return AssertionResult{Name: "unique-row-ids-within-page", Passed: true, Message: fmt.Sprintf("rows=%d", len(run.RowIDs))}
-}
-
-func validateFilterMatch(workload WorkloadDefinition, records []*model.PersistentRecord) AssertionResult {
-	for _, record := range records {
-		for key, expected := range workload.ResolvedFilterConditions() {
-			if !recordMatchesFilter(record, key, fmt.Sprint(expected)) {
-				return AssertionResult{Name: "filter-results-match-request", Passed: false, Message: fmt.Sprintf("attribute=%s expected=%v row=%s", key, expected, record.RowID)}
-			}
-		}
-	}
-	return AssertionResult{Name: "filter-results-match-request", Passed: true, Message: fmt.Sprintf("conditions=%v", workload.ResolvedFilterConditions())}
-}
-
-func validateTradeTimeWindow(records []*model.PersistentRecord, semantics workloadSemantics) AssertionResult {
-	for _, record := range records {
-		tradeTime, ok := record.Int64Items["tradeTime"]
-		if !ok {
-			return AssertionResult{Name: "tradeTime-window-match-request", Passed: false, Message: fmt.Sprintf("missing tradeTime row=%s", record.RowID)}
-		}
-		if semantics.TradeTimeStart > 0 && tradeTime < semantics.TradeTimeStart {
-			return AssertionResult{Name: "tradeTime-window-match-request", Passed: false, Message: fmt.Sprintf("row=%s tradeTime=%d start=%d", record.RowID, tradeTime, semantics.TradeTimeStart)}
-		}
-		if semantics.TradeTimeEnd > 0 && tradeTime > semantics.TradeTimeEnd {
-			return AssertionResult{Name: "tradeTime-window-match-request", Passed: false, Message: fmt.Sprintf("row=%s tradeTime=%d end=%d", record.RowID, tradeTime, semantics.TradeTimeEnd)}
-		}
-	}
-	return AssertionResult{Name: "tradeTime-window-match-request", Passed: true, Message: fmt.Sprintf("rows=%d start=%d end=%d", len(records), semantics.TradeTimeStart, semantics.TradeTimeEnd)}
-}
-
-func validateSortOrder(records []*model.PersistentRecord, attribute string, desc bool) AssertionResult {
-	values := make([]int64, 0, len(records))
-	for _, record := range records {
-		value, ok := record.Int64Items[attribute]
-		if !ok {
-			continue
-		}
-		values = append(values, value)
-	}
-	if len(values) <= 1 {
-		return AssertionResult{Name: "sorted-by-tradeTime-desc", Passed: true, Message: "insufficient comparable rows"}
-	}
-	if desc {
-		for i := 1; i < len(values); i++ {
-			if values[i] > values[i-1] {
-				return AssertionResult{Name: "sorted-by-tradeTime-desc", Passed: false, Message: fmt.Sprintf("index=%d prev=%d curr=%d", i, values[i-1], values[i])}
-			}
-		}
-	}
-	return AssertionResult{Name: "sorted-by-tradeTime-desc", Passed: true, Message: fmt.Sprintf("comparable_rows=%d", len(values))}
-}
-
-// benchmarkTradeFilterColumns and benchmarkTradeFilterEAVAttrs map trade filter
-// attributes to their storage layout (see schemas/trade_attributes.json):
-// service-path records are rebuilt via transform.ToPersistentRecord, which
-// stores column-bound attributes under entity_main column names and EAV-only
-// attributes as OtherAttributes entries.
-var benchmarkTradeFilterColumns = map[string]string{
-	"symbol": "text_01",
-	"region": "text_02",
-}
-
-var benchmarkTradeFilterEAVAttrs = map[string]int16{
-	"exchange":     8,
-	"orderChannel": 12,
-}
-
-func recordMatchesFilter(record *model.PersistentRecord, attribute, expected string) bool {
-	switch attribute {
-	case "symbol", "exchange", "region", "name", "orderChannel":
-		value, ok := benchmarkRecordTextValue(record, attribute)
-		return ok && value == expected
-	case "tradeType":
-		if value, ok := record.Int64Items[attribute]; ok {
-			return fmt.Sprintf("%d", value) == expected
-		}
-		if value, ok := record.Int16Items["smallint_01"]; ok {
-			return fmt.Sprintf("%d", value) == expected
-		}
-		return false
-	default:
-		// Fail closed: an unmapped filter attribute means the oracle cannot
-		// verify the row, and silently passing would mask filter regressions.
-		return false
-	}
-}
-
-// benchmarkRecordTextValue reads a benchmark text attribute from either record
-// shape: harness-path records carry attribute names directly, service-path
-// records carry the storage layout.
-func benchmarkRecordTextValue(record *model.PersistentRecord, attribute string) (string, bool) {
-	if value, ok := record.TextItems[attribute]; ok {
-		return value, true
-	}
-	if column, ok := benchmarkTradeFilterColumns[attribute]; ok {
-		if value, ok := record.TextItems[column]; ok {
-			return value, true
-		}
-	}
-	if attrID, ok := benchmarkTradeFilterEAVAttrs[attribute]; ok {
-		for _, eav := range record.OtherAttributes {
-			if eav.AttrID == attrID && eav.ValueText != nil {
-				return *eav.ValueText, true
-			}
-		}
-	}
-	return "", false
-}
-
-func persistentRecordIDs(records []*model.PersistentRecord) []string {
-	ids := make([]string, 0, len(records))
-	for _, record := range records {
-		ids = append(ids, record.RowID.String())
-	}
-	return ids
-}
-
-func hasRowIDOverlap(a, b []string) bool {
-	seen := make(map[string]struct{}, len(a))
-	for _, rowID := range a {
-		seen[rowID] = struct{}{}
-	}
-	for _, rowID := range b {
-		if _, ok := seen[rowID]; ok {
-			return true
-		}
-	}
-	return false
-}
-
-func countFailedAssertions(assertions []AssertionResult) int {
-	failed := 0
-	for _, assertion := range assertions {
-		if !assertion.Passed {
-			failed++
-		}
-	}
-	return failed
-}
-
-func extractPerTierMetrics(plan *model.ExecutionPlan) *PerTierMetrics {
-	if plan == nil {
-		return nil
-	}
-	metrics := &PerTierMetrics{Sources: len(plan.Sources)}
-	for _, src := range plan.Sources {
-		switch src.Engine {
-		case "postgres":
-			if src.ActualRows > 0 {
-				metrics.PGRows += src.ActualRows
-			}
-			if src.Reason == "dirty id set fetched" {
-				metrics.PGDirtyCount = src.ActualRows
-			}
-			if src.PredicatePushdown {
-				metrics.HasPushdown = true
-			}
-		case "duckdb":
-			if src.ActualRows > 0 {
-				metrics.DuckDBRows += src.ActualRows
-			}
-		}
-	}
-	if metrics.PGRows > 0 || metrics.DuckDBRows > 0 {
-		metrics.FinalRows = metrics.PGRows + metrics.DuckDBRows
-		if metrics.FinalRows > 0 {
-			metrics.PushdownRatio = float64(metrics.PGRows) / float64(metrics.FinalRows)
-		}
-	}
-	return metrics
-}
-
-func validatePushdownAssertions(run WorkloadRunResult) []AssertionResult {
-	if run.PerTier == nil {
-		return nil
-	}
-	var assertions []AssertionResult
-	pt := run.PerTier
-
-	if pt.Sources > 0 {
-		assertions = append(assertions, AssertionResult{
-			Name:    "pushdown-plan-sources-present",
-			Passed:  true,
-			Message: fmt.Sprintf("sources=%d", pt.Sources),
-		})
-	} else {
-		assertions = append(assertions, AssertionResult{
-			Name:    "pushdown-plan-sources-present",
-			Passed:  false,
-			Message: "execution plan has no sources",
-		})
-	}
-
-	if pt.PGRows >= 0 {
-		assertions = append(assertions, AssertionResult{
-			Name:    "pushdown-pg-rows-tracked",
-			Passed:  true,
-			Message: fmt.Sprintf("pg_rows=%d", pt.PGRows),
-		})
-	}
-
-	if pt.HasPushdown {
-		assertions = append(assertions, AssertionResult{
-			Name:    "pushdown-active",
-			Passed:  true,
-			Message: "at least one source has pushdown enabled",
-		})
-	}
-
-	if pt.PushdownRatio > 0 && pt.PushdownRatio <= 1 {
-		assertions = append(assertions, AssertionResult{
-			Name:    "pushdown-efficiency-reasonable",
-			Passed:  true,
-			Message: fmt.Sprintf("pushdown_ratio=%.3f", pt.PushdownRatio),
-		})
-	}
-
-	return assertions
-}
-
-func maxInt(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
 }
